+++ lib/tots_auth.dart
@@ -1,5 +1,8 @@
 library tots_auth;
 
+export 'src/entities/tots_user_entity.dart';
+export 'src/entities/tots_token_entity.dart';
 export 'src/tots_auth.dart';
 export 'src/http/tots_auth_interceptor.dart';
+export 'src/http/tots_auth_http_extension.dart';
 export 'src/http/tots_auth_http.dart';