+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.3
+
+* Fix export classes.
+
+## 0.0.2
+
+* Add user entity. Add login endpoint.
+
 ## 0.0.1
 
 * Initial version